port as unit16_t

+++ include/mdns_cpp/mdns.hpp
@@ -24,7 +24,7 @@ struct ServiceInfo {
   std::string instance_name;
   std::string host_name;
   std::vector<std::string> addresses;
-  int port = 0;
+  uint16_t port = 0;
   std::vector<std::pair<std::string, std::string>> txt_records;
   bool has_ptr = false, has_srv = false, has_a = false, has_txt = false;
 };
